+++ src/Synapse/Synapse/Commands/ManagementCommands/SparkPool/NewAzureSynapseSparkPool.cs
@@ -22,7 +22,6 @@ using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
 using Microsoft.Azure.Management.Synapse.Models;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using System.Collections;
-using System.IO;
 using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.Synapse
@@ -114,11 +113,6 @@ namespace Microsoft.Azure.Commands.Synapse
         [ValidateNotNullOrEmpty]
         public string SparkVersion { get; set; }
 
-        [Parameter(ValueFromPipelineByPropertyName = false, Mandatory = false,
-            HelpMessage = HelpMessages.LibraryRequirementsFilePath)]
-        [ValidateNotNullOrEmpty]
-        public string LibraryRequirementsFilePath { get; set; }
-
         [Parameter(Mandatory = false, HelpMessage = HelpMessages.AsJob)]
         public SwitchParameter AsJob { get; set; }
 
@@ -173,23 +167,11 @@ namespace Microsoft.Azure.Commands.Synapse
                 throw new AzPSResourceNotFoundCloudException(string.Format(Resources.WorkspaceDoesNotExist, this.WorkspaceName));
             }
 
-            LibraryRequirements libraryRequirements = null;
-            if (this.IsParameterBound(c => c.LibraryRequirementsFilePath))
-            {
-                var powerShellDestinationPath = SessionState.Path.GetUnresolvedProviderPathFromPSPath(LibraryRequirementsFilePath);
-
-                libraryRequirements = new LibraryRequirements
-                {
-                    Filename = Path.GetFileName(powerShellDestinationPath),
-                    Content = this.ReadFileAsText(powerShellDestinationPath),
-                };
-            }
-
             var createParams = new BigDataPoolResourceInfo
             {
                 Location = existingWorkspace.Location,
                 Tags = TagsConversionHelper.CreateTagDictionary(this.Tag, validate: true),
-                NodeCount = this.enableAutoScale ? (int?) null : this.NodeCount,
+                NodeCount = this.enableAutoScale ? (int?)null : this.NodeCount,
                 NodeSizeFamily = NodeSizeFamily.MemoryOptimized,
                 NodeSize = NodeSize,
                 AutoScale = !this.enableAutoScale ? new AutoScaleProperties { Enabled = false } : new AutoScaleProperties
@@ -203,8 +185,7 @@ namespace Microsoft.Azure.Commands.Synapse
                     Enabled = EnableAutoPause.IsPresent,
                     DelayInMinutes = AutoPauseDelayInMinute
                 },
-                SparkVersion = this.SparkVersion,
-                LibraryRequirements = libraryRequirements
+                SparkVersion = this.SparkVersion
             };
 
             if (this.ShouldProcess(this.Name, string.Format(Resources.CreatingSynapseSparkPool, this.ResourceGroupName, this.WorkspaceName, this.Name)))