+++ src/Synapse/Synapse/Models/PSSynapseSparkPool.cs
@@ -1,4 +1,4 @@
-﻿// ----------------------------------------------------------------------------------
+// ----------------------------------------------------------------------------------
 //
 // Copyright Microsoft Corporation
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,7 +12,10 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using Microsoft.Azure.Commands.Synapse.Models.WorkspacePackages;
 using Microsoft.Azure.Management.Synapse.Models;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Microsoft.Azure.Commands.Synapse.Models
 {
@@ -32,6 +35,12 @@ namespace Microsoft.Azure.Commands.Synapse.Models
             this.DefaultSparkLogFolder = sparkPool?.DefaultSparkLogFolder;
             this.NodeSize = sparkPool?.NodeSize;
             this.NodeSizeFamily = sparkPool?.NodeSizeFamily;
+            this.IsComputeIsolationEnabled = sparkPool?.IsComputeIsolationEnabled ?? false;
+            this.SessionLevelPackagesEnabled = sparkPool?.SessionLevelPackagesEnabled ?? false;
+            this.CacheSize = sparkPool?.CacheSize ?? 0;
+            this.DynamicExecutorAllocation = sparkPool?.DynamicExecutorAllocation != null ? new PSDynamicExecutorAllocation(sparkPool.DynamicExecutorAllocation) : null;
+            this.WorkspacePackages = sparkPool?.CustomLibraries?.Select(library => new PSSynapseWorkspacePackage(library)).ToList() ?? null;
+            this.LastSucceededTimestamp = sparkPool?.LastSucceededTimestamp;
         }
 
         /// <summary>
@@ -55,6 +64,26 @@ namespace Microsoft.Azure.Commands.Synapse.Models
         public PSAutoPauseProperties AutoPause { get; set; }
 
         /// <summary>
+        /// Gets or sets whether compute isolation is required or not.
+        /// </summary>
+        public bool IsComputeIsolationEnabled { get; set; }
+
+        /// <summary>
+        /// Gets or sets whether session level packages enabled.
+        /// </summary>
+        public bool SessionLevelPackagesEnabled { get; set; }
+
+        /// <summary>
+        /// Gets or sets the cache size
+        /// </summary>
+        public int CacheSize { get; set; }
+
+        /// <summary>
+        /// Gets or sets dynamic Executor Allocation
+        /// </summary>
+        public PSDynamicExecutorAllocation DynamicExecutorAllocation { get; set; }
+
+        /// <summary>
         /// Gets the Spark events folder
         /// </summary>
         public string SparkEventsFolder { get; set; }
@@ -68,6 +97,12 @@ namespace Microsoft.Azure.Commands.Synapse.Models
         /// Gets library version requirements
         /// </summary>
         public PSLibraryRequirements LibraryRequirements { get; set; }
+
+        /// <summary>
+        /// Gets or sets list of custom libraries/packages associated with the
+        /// spark pool.
+        /// </summary>
+        public IList<PSSynapseWorkspacePackage> WorkspacePackages { get; set; }
 
         /// <summary>
         /// Gets the Apache Spark version.
@@ -91,5 +126,10 @@ namespace Microsoft.Azure.Commands.Synapse.Models
         /// Possible values include: 'None', 'MemoryOptimized'
         /// </summary>
         public string NodeSizeFamily { get; set; }
+
+        /// <summary>
+        /// Gets the time when the Big Data pool was updated successfully.
+        /// </summary>
+        public System.DateTime? LastSucceededTimestamp { get; private set; }
     }
 }